requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.3
autopep8==1.4.4
billiard==3.3.0.23
celery==3.1.26.post2
cffi==1.13.2
cryptography==2.8
Django==3.0
django-celery==3.3.1
django-crispy-forms==1.8.1
django-rest-knox==4.1.0
djangorestframework==3.10.3
importlib-metadata==1.2.0
kombu==3.0.37
more-itertools==8.0.2
Pillow==6.2.1
pycodestyle==2.5.0
pycparser==2.19
pytz==2019.3
redis==3.2.0
six==1.13.0
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
